Introduce temporary long doubles for result.
Several of these require a conversion down to their target precision,
and directly using "=t" with a variable of the target precision tells
the compiler that the result already is of the right type. Which is
false in these cases.

Also, get rid of __exp2l, which was just needed for the assembler code,
and is not needed after moving to C.

+++ src/math/i386/log.c
@@ -6,5 +6,5 @@ double log(double x)
 	long double res;
 	__asm__("fldln2" : "=t"(ln2));
 	__asm__("fyl2x" : "=t"(res) : "0"(x), "u"(ln2) : "st(1)");
-	return res;
+	return (double)res;
 }
